refactor: remove additional anycompany from the package name
