+++ src/main/java/za/co/anycompany/AnycompanyApplication.java
@@ -1,8 +1,7 @@
-package za.co.anycompany.anycompany;
+package za.co.anycompany;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 //import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 //@EnableJpaRepositories("za.co.anycompany.anycompany.persistence.repo")
