+++ src/main/java/za/co/anycompany/datalayer/CustomerRepository.java
@@ -1,8 +1,8 @@
-package za.co.anycompany.anycompany.datalayer;
+package za.co.anycompany.datalayer;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import za.co.anycompany.anycompany.model.Customer;
+import za.co.anycompany.model.Customer;
 
 import java.sql.*;
 import java.util.ArrayList;
